Renaming files inside cli to match
